安全

+++ dem/module/view.py
@@ -16,7 +16,8 @@ def get_first_events(events):
             first_events = np.zeros((INPUT_HEIGHT, INPUT_WIDTH, 3)) # bgr
             first_events[:,:,0] = events[0,0,0] * 255# r 
             first_events[:,:,1] = events[0,0,1] * 255
-            return first_events
+            # print(first_events.max(), first_events.min())
+            return first_events.astype(int)
     elif events.dim() == 4:#[TCHW]
         events = events.to('cpu')
         if BOOL_DISTINGUISH_EVENT:
